Added better exception handling and feedback for the frontend
The Global exception handler was reworked to throw custom errors. The need for custom errors is so the services are decoupled as much as possible from the web logic (like http statuses and etc.) The rework will allow future exceptions to be handled with ease. Also added the 'remember me' functionality's setup for token life durations.